requirements.txt
requests>=2.28.0
PyYAML>=6.0
beautifulsoup4>=4.12.0
pillow>=10.0.0
python-dotenv>=0.21.0
